requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
prometheus-flask-exporter==0.22.4
prometheus-client==0.17.1